requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.2
boto3==1.26.106
botocore==1.29.106
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-WTF==1.1.1
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lru-dict==1.1.8
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.0.4
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parsimonious==0.8.1
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.2
